# packages/types/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Vex Protocol API",
        "version": "4.3.0",
        "description": "REST API for the Vex encrypted chat platform. Messages are serialized using msgpack.",
        "license": {
            "name": "AGPL-3.0-or-later"
        },
        "contact": {
            "name": "Vex Protocol",
            "url": "https://github.com/vex-protocol/vex-protocol/tree/master/packages/types"
        }
    },
    "servers": [
        {
            "url": "https://api.vex.wtf",
            "description": "Production"
        },
        {
            "url": "http://localhost:16777",
            "description": "Local development"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    },
    "paths": {
...
        "/user/{id}": {
            "get": {
                "operationId": "getUser",
                "summary": "Get user profile",
                "description": "Retrieve the public profile for a user by their ID.",
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile",
                        "content": {
                            "application/msgpack": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "lastSeen": {
                                            "type": "string",
                                            "description": "Last activity timestamp"
                                        },
                                        "userID": {
                                            "type": "string",
                                            "description": "Unique user identifier"
                                        },
                                        "username": {
                                            "type": "string",
                                            "description": "Display username"
                                        }
                                    },
                                    "required": [
                                        "lastSeen",
                                        "userID",
                                        "username"
                                    ],
                                    "description": "Public user profile"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/user/{id}/servers": {
            "get": {
                "operationId": "listUserServers",
                "summary": "List servers for a user",
                "description": "List all chat servers the given user is a member of.",
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server list",
                        "content": {
                            "application/msgpack": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "icon": {
                                                "type": "string",
                                                "description": "Server icon file ID"
                                            },
                                            "name": {
                                                "type": "string",
                                                "description": "Server display name"
                                            },
                                            "serverID": {
                                                "type": "string",
                                                "description": "Unique server identifier"
                                            }
                                        },
                                        "required": [
                                            "name",
                                            "serverID"
                                        ],
                                        "description": "Chat server"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/server/{id}": {
            "get": {
                "operationId": "getServer",
                "summary": "Get a server",
                "description": "Retrieve details for a chat server by its ID.",
                "tags": [
                    "servers"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server details",
                        "content": {
                            "application/msgpack": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "icon": {
                                            "type": "string",
                                            "description": "Server icon file ID"
                                        },
                                        "name": {
                                            "type": "string",
                                            "description": "Server display name"
                                        },
                                        "serverID": {
                                            "type": "string",
                                            "description": "Unique server identifier"
                                        }
                                    },
                                    "required": [
                                        "name",
                                        "serverID"
                                    ],
                                    "description": "Chat server"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "createServer",
                "summary": "Create a server",
                "description": "Create a new chat server. The path parameter is used as the server name.",
                "tags": [
                    "servers"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server created",
                        "content": {
                            "application/msgpack": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "icon": {
                                            "type": "string",
                                            "description": "Server icon file ID"
                                        },
                                        "name": {
                                            "type": "string",
                                            "description": "Server display name"
                                        },
                                        "serverID": {
                                            "type": "string",
                                            "description": "Unique server identifier"
                                        }
                                    },
                                    "required": [
                                        "name",
                                        "serverID"
                                    ],
                                    "description": "Chat server"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "deleteServer",
                "summary": "Delete a server",
                "description": "Permanently delete a chat server. Requires owner-level permissions.",
                "tags": [
                    "servers"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "name": "id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server deleted"
                    },
                    "401": {
                        "description": "Insufficient permissions"
                    }
                }
            }
        },
...
    },
...
}